chore: workspace id location
