+++ frontend/src/components/GeneralSetting/BrandingSetting.vue
@@ -13,30 +13,6 @@
     </div>
     <div class="flex-1 lg:px-4">
       <div class="mb-4 mt-4 lg:mt-0">
-        <label class="flex items-center gap-x-2">
-          <span class="font-medium">
-            {{ $t("settings.general.workspace.id") }}
-          </span>
-        </label>
-        <div class="mb-3 text-sm text-gray-400">
-          {{ $t("settings.general.workspace.id-description") }}
-        </div>
-        <div class="mb-4 flex space-x-2">
-          <NInput
-            ref="workspaceIdField"
-            class="mb-4 w-full"
-            readonly
-            :value="workspaceId"
-            @click="selectWorkspaceId"
-          />
-          <NButton
-            v-if="isSupported"
-            :disabled="!workspaceId"
-            @click="handleCopyId"
-          >
-            <heroicons-outline:clipboard-document class="w-4 h-4" />
-          </NButton>
-        </div>
         <span class="font-medium">
           {{ $t("settings.general.workspace.logo") }}
         </span>
@@ -104,9 +80,8 @@
 </template>
 
 <script lang="ts" setup>
-import { useClipboard } from "@vueuse/core";
-import { NButton, NInput, NPopconfirm, NTooltip } from "naive-ui";
-import { computed, reactive, watchEffect, ref } from "vue";
+import { NButton, NPopconfirm, NTooltip } from "naive-ui";
+import { computed, reactive, watchEffect } from "vue";
 import { useI18n } from "vue-i18n";
 import { featureToRef, pushNotification } from "@/store";
 import { useActuatorV1Store } from "@/store/modules/v1/actuator";
@@ -140,7 +115,6 @@ const convertFileToBase64 = (file: File) =>
 
 const settingV1Store = useSettingV1Store();
 const { t } = useI18n();
-const workspaceIdField = ref<HTMLInputElement | null>(null);
 
 const state = reactive<LocalState>({
   displayName: "",
@@ -157,31 +131,6 @@ watchEffect(() => {
 const valid = computed((): boolean => {
   return !!state.displayName || !!state.logoFile;
 });
-
-const workspaceId = computed(() => {
-  return (
-    settingV1Store.getSettingByName("bb.workspace.id")?.value?.stringValue ?? ""
-  );
-});
-
-const selectWorkspaceId = () => {
-  workspaceIdField.value?.select();
-};
-
-const { copy: copyTextToClipboard, isSupported } = useClipboard({
-  legacy: true,
-});
-
-const handleCopyId = () => {
-  selectWorkspaceId();
-  copyTextToClipboard(workspaceId.value).then(() => {
-    pushNotification({
-      module: "bytebase",
-      style: "SUCCESS",
-      title: t("common.copied"),
-    });
-  });
-};
 
 const allowDelete = computed(() => {
   return settingV1Store.brandingLogo !== "";
